Rename things to zeno_build
