+++ README.md
@@ -1,3 +1,3 @@
-# llm-compare
+# zeno-build
 
 Set up common LLM applications with evaluation
