+++ pyproject.toml
@@ -3,14 +3,14 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "llm_compare"
+name = "zeno_build"
 authors = [
     { name = "Alex Cabrera", email = "cabrera@cmu.edu" },
     { name = "Graham Neubig", email = "neubig@gmail.com" },
 ]
 description = "A library for comparing multiple llm-based systems."
 readme = "README.md"
-repository = "https://github.com/zeno-ml/llm-compare"
+repository = "https://github.com/zeno-ml/zeno-build"
 requires-python = ">=3.9"
 license = { file = "LICENSE" }
 classifiers = [
@@ -30,7 +30,7 @@ exclude = ["*_test.py"]
 only-packages = true
 
 [tool.hatch.build.targets.wheel]
-packages = ["llm_compare"]
+packages = ["zeno_build"]
 
 [tool.hatch.version]
-path = "llm_compare/version.py"
+path = "zeno_build/version.py"
