+++ zeno_build/prompts/chat_generate.py
@@ -1,28 +1,28 @@
-"""Generate from a textual prompt."""
+"""Tools to generate from prompts."""
 
 import asyncio
 
 import openai
 import tqdm
 
-from llm_compare.models import api_based_model, global_models
-from llm_compare.prompts.prompt_utils import replace_variables
+from zeno_build.models import api_based_model, global_models
+from zeno_build.prompts import chat_prompt
 
 
-async def generate_from_text_prompt(
+async def generate_from_chat_prompt(
     variables: list[dict[str, str]],
-    prompt_template: str,
+    prompt_template: chat_prompt.ChatMessages,
     model_config: api_based_model.ApiBasedModelConfig,
     temperature: float,
     max_tokens: int,
     top_p: float,
 ) -> list[str]:
-    """Generate from a textual prompt.
+    """Generate from a list of chat-style prompts.
 
     Args:
-        variables: The source set of variables to consume.
+        variables: The variables to be replaced in the prompt template.
         prompt_template: The template for the prompt.
-        model_config: Configuration of the model.
+        api_based_model_config: The API-based model configuration.
         temperature: The temperature to use.
         max_tokens: The maximum number of tokens to generate.
         top_p: The top p value to use.
@@ -38,7 +38,7 @@ async def generate_from_text_prompt(
         async_responses = [
             openai.Completion.acreate(
                 engine=model_config.model,
-                prompt=replace_variables(prompt_template, vars),
+                prompt=prompt_template.to_text_prompt(vars),
                 temperature=temperature,
                 max_tokens=max_tokens,
                 top_p=top_p,
@@ -51,12 +51,7 @@ async def generate_from_text_prompt(
         async_responses = [
             openai.ChatCompletion.acreate(
                 model=model_config.model,
-                messages=[
-                    {
-                        "role": "user",
-                        "content": replace_variables(prompt_template, vars),
-                    },
-                ],
+                messages=prompt_template.to_openai_chat_completion_messages(vars),
                 temperature=temperature,
                 max_tokens=max_tokens,
                 top_p=top_p,
@@ -71,8 +66,8 @@ async def generate_from_text_prompt(
         results = []
         for vars in tqdm.tqdm(variables, "Generating synchronously from Cohere"):
             try:
-                prompt = replace_variables(prompt_template, vars)
                 assert global_models.cohere_client is not None
+                prompt = prompt_template.to_text_prompt(vars)
                 response = global_models.cohere_client.generate(
                     model=model_config.model,
                     prompt=prompt,
@@ -87,4 +82,4 @@ async def generate_from_text_prompt(
                 results.append("")
         return results
     else:
-        raise ValueError("Unknown model_config.provider, but you can add your own!")
+        raise ValueError("Unknown provider, but you can add your own!")
