+++ zeno_build/prompts/text_generate.py
@@ -1,28 +1,28 @@
-"""Tools to generate from prompts."""
+"""Generate from a textual prompt."""
 
 import asyncio
 
 import openai
 import tqdm
 
-from llm_compare.models import api_based_model, global_models
-from llm_compare.prompts import chat_prompt
+from zeno_build.models import api_based_model, global_models
+from zeno_build.prompts.prompt_utils import replace_variables
 
 
-async def generate_from_chat_prompt(
+async def generate_from_text_prompt(
     variables: list[dict[str, str]],
-    prompt_template: chat_prompt.ChatMessages,
+    prompt_template: str,
     model_config: api_based_model.ApiBasedModelConfig,
     temperature: float,
     max_tokens: int,
     top_p: float,
 ) -> list[str]:
-    """Generate from a list of chat-style prompts.
+    """Generate from a textual prompt.
 
     Args:
-        variables: The variables to be replaced in the prompt template.
+        variables: The source set of variables to consume.
         prompt_template: The template for the prompt.
-        api_based_model_config: The API-based model configuration.
+        model_config: Configuration of the model.
         temperature: The temperature to use.
         max_tokens: The maximum number of tokens to generate.
         top_p: The top p value to use.
@@ -38,7 +38,7 @@ async def generate_from_chat_prompt(
         async_responses = [
             openai.Completion.acreate(
                 engine=model_config.model,
-                prompt=prompt_template.to_text_prompt(vars),
+                prompt=replace_variables(prompt_template, vars),
                 temperature=temperature,
                 max_tokens=max_tokens,
                 top_p=top_p,
@@ -51,7 +51,12 @@ async def generate_from_chat_prompt(
         async_responses = [
             openai.ChatCompletion.acreate(
                 model=model_config.model,
-                messages=prompt_template.to_openai_chat_completion_messages(vars),
+                messages=[
+                    {
+                        "role": "user",
+                        "content": replace_variables(prompt_template, vars),
+                    },
+                ],
                 temperature=temperature,
                 max_tokens=max_tokens,
                 top_p=top_p,
@@ -66,8 +71,8 @@ async def generate_from_chat_prompt(
         results = []
         for vars in tqdm.tqdm(variables, "Generating synchronously from Cohere"):
             try:
+                prompt = replace_variables(prompt_template, vars)
                 assert global_models.cohere_client is not None
-                prompt = prompt_template.to_text_prompt(vars)
                 response = global_models.cohere_client.generate(
                     model=model_config.model,
                     prompt=prompt,
@@ -82,4 +87,4 @@ async def generate_from_chat_prompt(
                 results.append("")
         return results
     else:
-        raise ValueError("Unknown provider, but you can add your own!")
+        raise ValueError("Unknown model_config.provider, but you can add your own!")
